backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.29
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
apscheduler>=3.10.4
PyMuPDF>=1.24.0
python-docx>=1.1.0
groq>=0.5.0
